VC_CRM/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.0.0,<2.0.0
python-dotenv==1.1.0
python-telegram-bot==22.0
playwright==1.51.0
google-api-python-client==2.66.0
google-auth==2.38.0
google-auth-oauthlib==1.1.0
pytesseract==0.3.10
Pillow>=10.0.0,<12.0.0
PyMuPDF==1.23.7
python-pptx==0.6.21
nest_asyncio==1.6.0
gspread==5.12.4
oauth2client==4.1.3
